Refactor from Long to Int where reasonable, lint, simplify aoc 2023 day 3

+++ src/main/kotlin/extensions/ListExtensions.kt
@@ -7,7 +7,7 @@ import datastructures.LongGrid
 
 fun List<Double>.toProduct(): Double = this.reduce { a, b -> a * b }
 
-fun List<Long>.toProduct(): Long = this.reduce { a, b -> a * b }
+fun List<Int>.toProduct(): Int = this.reduce { a, b -> a * b }
 
 fun List<String>.toLongGrid(defaultValue: Char): LongGrid =
     LongGrid(
